corrected extreme approximation error introduced by a cast to (size_t) (long double) Changed dataset interface accordingly
git-svn-id: https://projets-systeme.lip6.fr/svn/research/libddd/trunk/Sources/TOOLS_C/COMPONENTS/tmp_libddd/libDDD@36 ab02c66d-5113-476d-a8b6-74e7035bd63a

+++ src/DataSet.h
@@ -30,7 +30,7 @@ class DataSet {
   // compares two sets for equality
   virtual bool set_equal(const DataSet & b) const =0;
   // returns the size (number of elements) in a set
-  virtual size_t set_size() const = 0;
+  virtual long double set_size() const = 0;
   // returns a hash function, used in the SDD hash function computation
   virtual size_t set_hash() const =0;
   // returns a formatted string description of the set
